Remove dependency on num
ocplib-simplex is written as a functor to allow for dependency
injection, so there is no good reason to depend on Num except to run the
tests.

Moreover, the tests are written using Num, which has been deprecated for
a long time, so this patch also switches them to use ZArith instead.